Combined index and class file for editEntry
