--- modules/battery_manager/js/editEntryIndex.js
@@ -1,2 +1,0 @@
-!function(modules){function __webpack_require__(moduleId){if(installedModules[moduleId])return installedModules[moduleId].exports;var module=installedModules[moduleId]={exports:{},id:moduleId,loaded:!1};return modules[moduleId].call(module.exports,module,module.exports,__webpack_require__),module.loaded=!0,module.exports}var installedModules={};return __webpack_require__.m=modules,__webpack_require__.c=installedModules,__webpack_require__.p="",__webpack_require__(0)}({0:function(module,exports,__webpack_require__){"use strict";function _interopRequireDefault(obj){return obj&&obj.__esModule?obj:{default:obj}}var _editEntry=__webpack_require__(5),_editEntry2=_interopRequireDefault(_editEntry),args=QueryString.get(document.currentScript.src);$(function(){var batteryManagerEditForm=React.createElement("div",{className:"page-edit-entry"},React.createElement("div",{className:"row"},React.createElement("div",{className:"col-md-9 col-lg-7"},React.createElement(_editEntry2.default,{DataURL:loris.BaseURL+"/battery_manager/ajax/get_form_data.php?form=edit&ID="+args.id,checkForDuplicate:loris.BaseURL+"/battery_manager/ajax/add_or_edit_entry.php?action=checkForDuplicate",activate:loris.BaseURL+"/battery_manager/ajax/change_active_status.php?action=activate",deactivate:loris.BaseURL+"/battery_manager/ajax/change_active_status.php?action=deactivate",edit:loris.BaseURL+"/battery_manager/ajax/add_or_edit_entry.php?action=edit"}))));ReactDOM.render(batteryManagerEditForm,document.getElementById("lorisworkspace"))})},5:function(module,exports){"use strict";function _classCallCheck(instance,Constructor){if(!(instance instanceof Constructor))throw new TypeError("Cannot call a class as a function")}function _possibleConstructorReturn(self,call){if(!self)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return!call||"object"!=typeof call&&"function"!=typeof call?self:call}function _inherits(subClass,superClass){if("function"!=typeof superClass&&null!==superClass)throw new TypeError("Super expression must either be null or a function, not "+typeof superClass);subClass.prototype=Object.create(superClass&&superClass.prototype,{constructor:{value:subClass,enumerable:!1,writable:!0,configurable:!0}}),superClass&&(Object.setPrototypeOf?Object.setPrototypeOf(subClass,superClass):subClass.__proto__=superClass)}Object.defineProperty(exports,"__esModule",{value:!0});var _createClass=function(){function defineProperties(target,props){for(var i=0;i<props.length;i++){var descriptor=props[i];descriptor.enumerable=descriptor.enumerable||!1,descriptor.configurable=!0,"value"in descriptor&&(descriptor.writable=!0),Object.defineProperty(target,descriptor.key,descriptor)}}return function(Constructor,protoProps,staticProps){return protoProps&&defineProperties(Constructor.prototype,protoProps),staticProps&&defineProperties(Constructor,staticProps),Constructor}}(),BatteryManagerEditForm=function(_React$Component){function BatteryManagerEditForm(props){_classCallCheck(this,BatteryManagerEditForm);var _this=_possibleConstructorReturn(this,(BatteryManagerEditForm.__proto__||Object.getPrototypeOf(BatteryManagerEditForm)).call(this,props));return _this.state={Data:{},formData:{},isLoaded:!1,loadedData:0},_this.setFormData=_this.setFormData.bind(_this),_this.handleEdit=_this.handleEdit.bind(_this),_this.isDuplicate=_this.isDuplicate.bind(_this),_this.giveOptions=_this.giveOptions.bind(_this),_this.activateEntry=_this.activateEntry.bind(_this),_this.deactivateEntry=_this.deactivateEntry.bind(_this),_this.editEntry=_this.editEntry.bind(_this),_this}return _inherits(BatteryManagerEditForm,_React$Component),_createClass(BatteryManagerEditForm,[{key:"componentDidMount",value:function(){var self=this;$.ajax(this.props.DataURL,{dataType:"json",success:function(data){var formData={id:data.batteryData.ID,instrument:data.batteryData.Test_name,ageMinDays:data.batteryData.AgeMinDays,ageMaxDays:data.batteryData.AgeMaxDays,stage:data.batteryData.Stage,subproject:data.batteryData.SubprojectID,visitLabel:data.batteryData.Visit_label,forSite:data.batteryData.CenterID,firstVisit:data.batteryData.firstVisit,instrumentOrder:data.batteryData.instr_order,active:data.batteryData.Active};self.setState({Data:data,isLoaded:!0,batteryData:data.batteryData,formData:formData})},error:function(_error,errorCode,errorMsg){console.error(_error,errorCode,errorMsg),self.setState({error:"An error occurred when loading the form!"})}})}},{key:"render",value:function(){if(void 0!==this.state.error)return React.createElement("div",{className:"alert alert-danger text-center"},React.createElement("strong",null,this.state.error));if(!this.state.isLoaded)return React.createElement("button",{className:"btn-info has-spinner"},"Loading",React.createElement("span",{className:"glyphicon glyphicon-refresh glyphicon-refresh-animate"}));var helpText=React.createElement("span",null,"You can activate or deactivate an entry by changing the Active value in the form.",React.createElement("br",null),React.createElement("br",null),"You cannot edit an entry to have the same values as another entry.",React.createElement("br",null),"If the edited entry has the same values but a different Active value as another entry,",React.createElement("br",null),"you can activate or deactivate the other entry based on the change you made.",React.createElement("br",null),"In this case, the entry that was originally selected for editing will not be edited.",React.createElement("br",null));return React.createElement("div",null,React.createElement(FormElement,{name:"batteryEdit",onSubmit:this.handleEdit,ref:"form"},React.createElement("h3",null,"Edit entry in Test Battery"),React.createElement("br",null),React.createElement(StaticElement,{label:"Note",text:helpText}),React.createElement(SelectElement,{name:"instrument",label:"Instrument",options:this.state.Data.instruments,onUserInput:this.setFormData,ref:"instrument",required:!0,value:this.state.formData.instrument}),React.createElement(NumericElement,{name:"ageMinDays",label:"Minimum age (days)",onUserInput:this.setFormData,ref:"ageMinDays",required:!0,min:"0",max:"99999",value:this.state.formData.ageMinDays}),React.createElement(NumericElement,{name:"ageMaxDays",label:"Maximum age (days)",onUserInput:this.setFormData,ref:"ageMaxDays",required:!0,min:"0",max:"99999",value:this.state.formData.ageMaxDays}),React.createElement(SelectElement,{name:"stage",label:"Stage",options:this.state.Data.stages,onUserInput:this.setFormData,ref:"stage",hasError:!1,required:!0,value:this.state.formData.stage}),React.createElement(SelectElement,{name:"subproject",label:"Subproject",options:this.state.Data.subprojects,onUserInput:this.setFormData,ref:"subproject",hasError:!1,required:!1,value:this.state.formData.subproject}),React.createElement(SelectElement,{name:"visitLabel",label:"Visit Label",options:this.state.Data.visits,onUserInput:this.setFormData,ref:"visitLabel",required:!1,value:this.state.formData.visitLabel}),React.createElement(SelectElement,{name:"forSite",label:"Site",options:this.state.Data.sites,onUserInput:this.setFormData,ref:"forSite",required:!1,value:this.state.formData.forSite}),React.createElement(SelectElement,{name:"firstVisit",label:"First Visit",options:this.state.Data.firstVisits,onUserInput:this.setFormData,ref:"firstVisit",required:!1,value:this.state.formData.firstVisit}),React.createElement(NumericElement,{name:"instrumentOrder",label:"Instrument Order",onUserInput:this.setFormData,ref:"instrumentOrder",required:!1,min:"0",max:"127",value:this.state.formData.instrumentOrder}),React.createElement(SelectElement,{name:"active",label:"Active",emptyOption:!1,options:this.state.Data.active,onUserInput:this.setFormData,ref:"active",value:this.state.formData.active}),React.createElement(ButtonElement,{label:"Edit entry"})))}},{key:"handleEdit",value:function(e){e.preventDefault();var formData=this.state.formData,formObj=new FormData;for(var key in formData)null!==formData[key]&&formObj.append(key,formData[key]);this.isDuplicate(formObj,this.giveOptions)}},{key:"isDuplicate",value:function(formObj,callback){$.ajax({type:"POST",url:this.props.checkForDuplicate,data:formObj,cache:!1,contentType:!1,processData:!1,success:callback,error:function(err){console.error(err),swal("Could not check","","error")}})}},{key:"giveOptions",value:function(duplicateEntry){var editedEntry=this.state.formData;if(Object.keys(duplicateEntry).length>0){var duplicateEntryJSON=JSON.parse(duplicateEntry),entryID=duplicateEntryJSON.ID,idObj=new FormData;if(idObj.append("ID",entryID),editedEntry.active===duplicateEntryJSON.Active){var errorMessage="The changes you made are identical to another entry in the table.";editedEntry.id===duplicateEntryJSON.ID&&(errorMessage="You did not make any changes to the current entry."),swal({title:"Duplicate entry!",text:errorMessage,type:"error"})}else if("Y"===editedEntry.active&&"N"===duplicateEntryJSON.Active){var warningMessage="The changes you made are identical to another entry in the table, with the exception of the Active status.\n Would you like to activate the other entry?\n Note: No changes will be made to the current entry.";editedEntry.id===duplicateEntryJSON.ID&&(warningMessage="You did not make any changes to the current entry except for the Active status.\n Would you like to activate this entry?"),swal({title:"Activate entry?",text:warningMessage,type:"warning",showCancelButton:!0,confirmButtonText:"Yes",cancelButtonText:"Cancel",closeOnConfirm:!1},function(){this.activateEntry(idObj)}.bind(this))}else if("N"===editedEntry.active&&"Y"===duplicateEntryJSON.Active){var _warningMessage="The changes you made are identical to another entry in the table, with the exception of the Active status.\n Would you like to deactivate the other entry?\n Note: No changes will be made to the current entry.";editedEntry.id===duplicateEntryJSON.ID&&(_warningMessage="You did not make any changes to the current entry except for the Active status.\n Would you like to deactivate this entry?"),swal({title:"Deactivate entry?",text:_warningMessage,type:"warning",showCancelButton:!0,confirmButtonText:"Yes",cancelButtonText:"Cancel",closeOnConfirm:!1},function(){this.deactivateEntry(idObj)}.bind(this))}}else this.editEntry()}},{key:"activateEntry",value:function(idObj){$.ajax({type:"POST",url:this.props.activate,data:idObj,cache:!1,contentType:!1,processData:!1}).done(function(data){swal({title:"Activated!",type:"success"},function(){window.location.assign(loris.BaseURL+"/battery_manager/")})}).error(function(data){swal("Could not activate entry","","error")})}},{key:"deactivateEntry",value:function(idObj){$.ajax({type:"POST",url:this.props.deactivate,data:idObj,cache:!1,contentType:!1,processData:!1}).done(function(data){swal({title:"Deactivated!",type:"success"},function(){window.location.assign(loris.BaseURL+"/battery_manager/")})}).error(function(data){swal("Could not deactivate entry","","error")})}},{key:"editEntry",value:function(){var formData=this.state.formData,formObj=new FormData;for(var key in formData)formData.hasOwnProperty(key)&&formObj.append(key,formData[key]);$.ajax({type:"POST",url:this.props.edit,data:formObj,cache:!1,contentType:!1,processData:!1,success:function(data){this.setState({formData:{}}),swal({title:"Edit Successful!",type:"success"},function(){window.location.assign(loris.BaseURL+"/battery_manager/")})}.bind(this),error:function(err){console.error(err),swal("Could not edit!",err.responseText,"error")}})}},{key:"setFormData",value:function(formElement,value){var formData=this.state.formData;""===value?formData[formElement]=null:formData[formElement]=value,this.setState({formData:formData})}}]),BatteryManagerEditForm}(React.Component);BatteryManagerEditForm.propTypes={DataURL:React.PropTypes.string.isRequired,activate:React.PropTypes.string.isRequired,deactivate:React.PropTypes.string.isRequired,checkForDuplicate:React.PropTypes.string.isRequired,edit:React.PropTypes.string.isRequired},exports.default=BatteryManagerEditForm}});
-//# sourceMappingURL=editEntryIndex.js.map
